+++ tests/test_container.py
@@ -1,20 +1,19 @@
+import asyncio
+
 import aioxmpp
-import pytest
 from asynctest import MagicMock, CoroutineMock
 
-from spade.behaviour import OneShotBehaviour
+from spade.behaviour import OneShotBehaviour, CyclicBehaviour
 from spade.container import Container
 from spade.message import Message
-from tests.utils import make_connected_agent
-
-from tests.utils import run_around_tests
+from .factories import MockedAgentFactory
 
 
 def test_use_container():
     container = Container()
     container.reset()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
 
     assert agent.container == Container()
 
@@ -54,7 +53,7 @@ def test_send_message_with_container():
 
     fake_receiver_agent.dispatch = MagicMock()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -86,7 +85,7 @@ def test_send_message_to_outer_with_container():
     container = Container()
     container.reset()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.start(auto_register=False)
 
     behaviour = SendBehaviour()
@@ -109,8 +108,8 @@ def test_unregister():
     container = Container()
     container.reset()
 
-    agent = make_connected_agent()
-    agent2 = make_connected_agent(jid="agent2@server")
+    agent = MockedAgentFactory()
+    agent2 = MockedAgentFactory(jid="agent2@server")
 
     assert container.has_agent(str(agent.jid))
     assert container.has_agent(str(agent2.jid))
@@ -119,3 +118,25 @@ def test_unregister():
 
     assert not container.has_agent(str(agent.jid))
     assert container.has_agent(str(agent2.jid))
+
+
+def test_cancel_tasks():
+    agent = MockedAgentFactory()
+
+    class Behav(CyclicBehaviour):
+        async def run(self):
+            await asyncio.sleep(100)
+            self.has_finished = True
+
+    behav = Behav()
+    behav.has_finished = False
+    agent.add_behaviour(behaviour=behav)
+    future = agent.start()
+    future.result()
+
+    assert not behav.has_finished
+
+    container = Container()
+    container.stop()
+
+    assert not behav.has_finished