+++ tests/test_web.py
@@ -15,13 +15,11 @@ from testfixtures import LogCapture
 from spade.agent import Agent
 from spade.behaviour import OneShotBehaviour, CyclicBehaviour
 from spade.message import Message
-from tests.utils import make_connected_agent, make_presence_connected_agent
-
-from tests.utils import run_around_tests
+from .factories import MockedAgentFactory, MockedPresenceAgentFactory
 
 
 def test_web():
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
     agent.web.start(port=10000)
@@ -82,7 +80,7 @@ def test_add_template_path():
 
 
 async def test_check_server(test_client):
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -100,11 +98,11 @@ async def test_check_server(test_client):
 
     assert sel.css("ul.products-list > li").getall() == []
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_request_home(test_client):
-    agent = make_connected_agent("jid@server", "password")
+    agent = MockedAgentFactory(jid="jid@server", password="password")
     future = agent.start(auto_register=False)
     future.result()
     agent.web.setup_routes()
@@ -120,7 +118,7 @@ async def test_request_home(test_client):
 
     assert sel.css("ul.products-list > li").getall() == []
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_get_messages(test_client):
@@ -140,7 +138,7 @@ async def test_get_messages(test_client):
 
     assert len(sel.css("ul.timeline > li").getall()) == 6  # num messages + end clock
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_get_behaviour(test_client):
@@ -166,7 +164,7 @@ async def test_get_behaviour(test_client):
         sel.css("section.content-header > h1::text").get().strip()
         == "OneShotBehaviour/EmptyOneShotBehaviour"
     )
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_kill_behaviour(test_client):
@@ -185,11 +183,11 @@ async def test_kill_behaviour(test_client):
 
     assert behaviour.is_killed()
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_get_agent(test_client):
-    agent = make_presence_connected_agent("jid@server", "password")
+    agent = MockedPresenceAgentFactory(jid="jid@server", password="password")
     future = agent.start(auto_register=False)
     future.result()
 
@@ -208,11 +206,11 @@ async def test_get_agent(test_client):
 
     assert sel.css("section.content-header > h1::text").get().strip() == jid
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_unsubscribe_agent(test_client):
-    agent = make_presence_connected_agent()
+    agent = MockedPresenceAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
     agent.client.enqueue = Mock()
@@ -236,11 +234,11 @@ async def test_unsubscribe_agent(test_client):
     assert arg.to == jid_.bare()
     assert arg.type_ == PresenceType.UNSUBSCRIBE
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_send_agent(test_client):
-    agent = make_presence_connected_agent()
+    agent = MockedPresenceAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
     agent.stream = MagicMock()
@@ -263,7 +261,7 @@ async def test_send_agent(test_client):
     assert sent[1].sent
     assert sent[1].body == "Hello World"
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_find_behaviour():
@@ -271,14 +269,14 @@ async def test_find_behaviour():
         async def run(self):
             pass
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = EmptyOneShotBehaviour()
     agent.add_behaviour(behaviour)
     found_behaviour = agent.web.find_behaviour("OneShotBehaviour/EmptyOneShotBehaviour")
 
     assert found_behaviour == behaviour
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_find_behaviour_fail():
@@ -287,7 +285,7 @@ async def test_find_behaviour_fail():
 
     assert found_behaviour is None
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_add_get(test_client):
@@ -303,7 +301,7 @@ async def test_add_get(test_client):
     sel = Selector(text=response)
     assert sel.css("h1::text").get().strip() == "42"
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_add_get_raw(test_client):
@@ -323,7 +321,7 @@ async def test_add_get_raw(test_client):
 
     assert response == "Hello Raw Get"
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_add_post(test_client):
@@ -344,7 +342,7 @@ async def test_add_post(test_client):
     sel = Selector(text=response)
     assert sel.css("h1::text").get() == "1024"
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_add_post_raw(test_client):
@@ -364,7 +362,7 @@ async def test_add_post_raw(test_client):
 
     assert response == "Hello Raw Post Number=1024"
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_stop(test_client):
@@ -415,7 +413,7 @@ async def test_add_get_json(test_client):
     data = await response.json()
     assert data["number"] == 42
 
-    agent.stop()
+    agent.stop().result()
 
 
 async def test_add_post_json(test_client):
@@ -436,4 +434,4 @@ async def test_add_post_json(test_client):
     data = await response.json()
     assert data["number"] == 1024
 
-    agent.stop()
+    agent.stop().result()